changes in student and college entity.

+++ src/main/java/com/sih/ais/dao/CollegeRepo.java
@@ -1,8 +1,7 @@
 package com.sih.ais.dao;
 
-
 import com.sih.ais.Entity.dto.College;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CollegeRepo extends JpaRepository<College,Long> {
+public interface CollegeRepo extends JpaRepository<College,Integer> {
 }
